*: isolate logutil subpackage
