+++ server/server.go
@@ -98,6 +98,7 @@ func NewServer(cfg *Config) (*Server, error) {
 	}
 
 	go systimemon.StartMonitor(time.Now, func() {
+		log.Errorf("system time jumps backward")
 		timeJumpBackCounter.Inc()
 	})
 	return s, nil
